Running server and LLM in different processes (#110)
* Putting server and LLM on different processes

* result queues cleanup

* Syncing server start with LLM init

* Renamed model to protocol to match the original endpoint

* Removing tuning parameters that are no longer used

* process spawn method from env
